Update PayloadFactory
Updated return values to integer for iat,exp,nbf claims
Closes #165

+++ src/PayloadFactory.php
@@ -155,7 +155,7 @@ class PayloadFactory
      */
     public function iat()
     {
-        return Utils::now()->format('U');
+        return (int) Utils::now()->format('U');
     }
 
     /**
@@ -165,7 +165,7 @@ class PayloadFactory
      */
     public function exp()
     {
-        return Utils::now()->addMinutes($this->ttl)->format('U');
+        return (int) Utils::now()->addMinutes($this->ttl)->format('U');
     }
 
     /**
@@ -175,7 +175,7 @@ class PayloadFactory
      */
     public function nbf()
     {
-        return Utils::now()->format('U');
+        return (int) Utils::now()->format('U');
     }
 
     /**
